fix: refactor coverage gutter formatter (#585)
e234ae6 feat: add back light theme status decorations
0c930ef refactor: use common functions from decorations
e916d74 fix: use prepared svg icon in coverage formatter
