+++ src/decorations/StateDecorations.ts
@@ -1,5 +1,3 @@
-import * as path from 'path';
-import * as fs from 'fs';
 import {
   window,
   OverviewRulerLane,
@@ -12,58 +10,38 @@ import passingIcon from 'vscode-codicons/src/icons/check.svg';
 import failingIcon from 'vscode-codicons/src/icons/chrome-close.svg';
 import skipIcon from 'vscode-codicons/src/icons/debug-step-over.svg';
 import unknownIcon from 'vscode-codicons/src/icons/question.svg';
+import prepareIcon from './prepareIcon';
 
-export class Decorations {
-  private static ICONS_PATH = path.join('out', 'icons');
-
-  private context: ExtensionContext;
-
+export class StateDecorations {
   public passing: TextEditorDecorationType;
   public failing: TextEditorDecorationType;
   public skip: TextEditorDecorationType;
   public unknown: TextEditorDecorationType;
 
-  constructor(context) {
-    this.context = context;
-
-    this.passing = this.createStateDecoration([['passing', passingIcon, '#35A15E'], 'green']);
-    this.failing = this.createStateDecoration([['failing', failingIcon, '#D6443C'], 'red']);
-    this.skip = this.createStateDecoration([['skip', skipIcon, '#fed37f'], 'yellow']);
+  constructor(context: ExtensionContext) {
+    this.passing = this.createStateDecoration([
+      prepareIcon(context, 'passing', passingIcon, '#35A15E'),
+      'green',
+    ]);
+    this.failing = this.createStateDecoration([
+      prepareIcon(context, 'failing', failingIcon, '#D6443C'),
+      'red',
+    ]);
+    this.skip = this.createStateDecoration([
+      prepareIcon(context, 'skip', skipIcon, '#fed37f'),
+      'yellow',
+    ]);
     this.unknown = this.createStateDecoration(
-      [['unknown', unknownIcon, '#BBBBBB'], 'darkgrey'],
-      [['unknown-light', unknownIcon, '#555555']]
+      [prepareIcon(context, 'unknown', unknownIcon, '#BBBBBB'), 'darkgrey'],
+      [prepareIcon(context, 'unknown-light', unknownIcon, '#555555')]
     );
   }
 
-  private resolvePath(...args: string[]): string {
-    return this.context.asAbsolutePath(path.join(...args));
-  }
-
-  private prepareIcon(state: string, source: string, color?: string): string {
-    const resultIconPath = this.resolvePath(Decorations.ICONS_PATH, `${state}.svg`);
-    let result = source.toString();
-
-    if (color !== undefined) {
-      result = result.replace('fill="currentColor"', `fill="${color}"`);
-    }
-
-    if (!fs.existsSync(resultIconPath) || fs.readFileSync(resultIconPath).toString() !== result) {
-      if (!fs.existsSync(this.resolvePath(Decorations.ICONS_PATH))) {
-        fs.mkdirSync(this.resolvePath(Decorations.ICONS_PATH));
-      }
-
-      fs.writeFileSync(resultIconPath, result);
-    }
-
-    return resultIconPath;
-  }
-
   private createStateDecoration(
-    dark: /* default */ [Parameters<Decorations['prepareIcon']>, string?],
-    light?: /* optional overrides */ [Parameters<Decorations['prepareIcon']>, string?]
+    dark: /* default */ [string, string?],
+    light?: /* optional overrides */ [string, string?]
   ): TextEditorDecorationType {
-    const [iconOptions, overviewRulerColor] = dark;
-    const icon = this.prepareIcon(...iconOptions);
+    const [icon, overviewRulerColor] = dark;
 
     const options: DecorationRenderOptions = {
       gutterIconPath: icon,
@@ -74,7 +52,7 @@ export class Decorations {
         gutterIconPath: icon,
       },
       light: {
-        gutterIconPath: light !== undefined ? this.prepareIcon(...light[0]) : icon,
+        gutterIconPath: light !== undefined ? light[0] : icon,
       },
     };
